chat

+++ app/Http/Controllers/Admin/ProfileController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers\Admin;
 
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use App\Models\Admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Validator;
 
 class ProfileController extends Controller
 {
@@ -17,43 +19,65 @@ class ProfileController extends Controller
 
     public function edit($id)
     {
-        $admin = Admin::query()->find($id);
+
+        if(Auth::user()->super_admin == 1){
+            $admin = Admin::query()->find($id);
+        }else{
+            $admin = Admin::query()->find(Auth::user()->id);
+        }
         return view('admin.profile.profile')->with(['admin' => $admin]);
     }
 
-
-    public function update(Request $request, $id)
+    public function updateAvatar(Request $request)
     {
-        $admin = Admin::findOrFail($id);
-        $admin->name = $request->name;
-        $admin->surname = $request->surname;
 
+        $validation = Validator::make($request->all(), [
+            'avatar' => 'required|image',
+        ])->validate();
 
-        if ($request->avatar != null) {
-
-//            $path = Storage::disk('local')->put('avatars', $request->avatar);
-//            dd($request->avatar);
-//            dd($path);
-            $path = file_url('app/avatars'.$request->avatar);
-            $admin->avatar = $path;
+//        dd($validation->fails());
+        if ($validation->fails())
+        {
+            return redirect()->back()->withErrors($validation->errors());
         }
 
+        $admin = Admin::findOrFail(Auth::user()->id);
+        $path = $request->file('avatar')->store('public/avatars');
+        $admin->avatar = file_url($path);
+        $admin->save();
+
+        return redirect()->back();
+    }
+
+    public function updateInfo(Request $request)
+    {
+        $admin = Admin::findOrFail(Auth::user()->id);
+        $admin->name = $request->name;
+        $admin->surname = $request->surname;
 
         if ($request->email != null) {
             $admin->email = $request->email;
         }
-        if ($request->password != null) {
-            $admin->password = Hash::make($request->password);
-        }
+
         $admin->save();
         return redirect()->back();
     }
 
+    public function updatePassword(Request $request)
+    {
+        $admin = Admin::findOrFail(Auth::user()->id);
+        $admin->password = bcrypt($request->password);
+        $admin->save();
+        return redirect()->back();
+    }
+
+
     public function destroy($id)
     {
-        $target = Admin::findOrFail($id);
-        $target->delete();
-
+        if(Auth::user()->super_admin == 1){
+            $target = Admin::findOrFail($id);
+            $target->delete();
+        }
         return redirect()->back();
     }
